cleaned ques_images page
