+++ project/app_utils/llm_utils.py
@@ -16,6 +16,12 @@ from langchain_community.llms import HuggingFaceHub
 from docx import Document
 import os
 
+import base64
+from PIL import Image
+
+
+
+
 load_dotenv()
 
 
@@ -77,6 +83,10 @@ def get_vectorstore(text_chunks:list):
     """
     Function returns vectorstore from FAISS from list of text chunks
 
+    For embedding, we use OpenAI embediing model( default)
+    
+    Then, we use FAISS to create a vector store using the embedding model. ie. vector store where our text chunks are represented as numbers.
+
     Args:
         text_chunks (list): _description_
 
@@ -127,3 +137,28 @@ def handle_user_input(user_question, conversation):
             st.write("**Question:**", message.content)
         else:
             st.write("**Response:**", message.content)
+
+
+def encode_image(uploaded_image):
+    """Encodes an uploaded image to base64
+
+    Args:
+        uploaded_image (UploadedFile): The uploaded image file from Streamlit.
+
+    Returns:
+        str: Base64 encoded string of the image
+    """
+    image_bytes = uploaded_image.read()
+    return base64.b64encode(image_bytes).decode('utf-8')
+    
+
+
+def display_uploaded_image(uploaded_image):
+    """
+    Displays the uploaded image from the user
+
+    Args:
+        uploaded_image (_type_): The uploaded image file from Streamlit.
+    """
+    image_to_display = Image.open(uploaded_image)
+    return image_to_display
